backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
pytesseract==0.3.10
easyocr==1.7.1
numpy==1.26.3
Pillow==10.2.0
openai==1.12.0
python-dotenv==1.0.0
